e2classaverage.py up and running, some aligner debug

+++ libEM/reconstructor_tools.cpp
@@ -160,7 +160,7 @@ bool InterpolatedFRC::continue_frc_calc_functoid(const float& xx, const float& y
 	int zt = z0 - nz/2;
 	
 	int radius = (int) (x_scale*x_scale*floor(xx)*floor(xx) + y_scale*y_scale*yt*yt + z_scale*z_scale*zt*zt);
-	radius = static_cast<int>(sqrtf(radius)*bin);
+	radius = static_cast<int>(sqrt(radius)*bin);
 	
 	if ( radius > (size-1) )
 	{
@@ -243,7 +243,7 @@ bool InterpolatedFRC::continue_frc_calc5(const float& xx, const float& yy, const
 	int zt = z0 - nz/2;
 
 	int radius = (int) (x_scale*x_scale*x0*x0 + y_scale*y_scale*yt*yt + z_scale*z_scale*zt*zt);
-	radius = static_cast<int>(sqrtf(radius)*bin);
+	radius = static_cast<int>(sqrt(radius)*bin);
 
 	if ( radius > (size-1) )
 	{
@@ -377,7 +377,7 @@ bool InterpolatedFRC::continue_frc_calc2(const float& xx, const float& yy, const
 
 	int radius = (int) (x_scale*x_scale*x0*x0 + y_scale*y_scale*yt*yt + z_scale*z_scale*zt*zt);
 // 	int radius =  x0*x0 + y_scale*y_scale*yt*yt + z_scale*z_scale*zt*zt;
-	radius = static_cast<int>(sqrtf(radius)*bin);
+	radius = static_cast<int>(sqrt(radius)*bin);
 
 	if ( radius > (size-1) )
 	{
@@ -453,7 +453,7 @@ bool InterpolatedFRC::continue_frc_calc1(const float& xx, const float& yy, const
 	int zt = z0 - nz/2;
 	
 	int radius = (int) (x_scale*x_scale* floor(xx)*floor(xx) + y_scale*y_scale*yt*yt + z_scale*z_scale*zt*zt);
-	radius = static_cast<int>(sqrtf(radius)*bin);
+	radius = static_cast<int>(sqrt(radius)*bin);
 	
 	if ( radius > (size-1) ) return false;
 
@@ -505,7 +505,7 @@ QualityScores InterpolatedFRC::finish(const unsigned int num_particles)
 		if ( frc_norm_rdata[i] == 0 || frc_norm_dt[i] == 0 )
 			frc[i] = 0;
 		else
-			frc[i] /= sqrtf(frc_norm_rdata[i]*frc_norm_dt[i]);
+			frc[i] /= sqrt(frc_norm_rdata[i]*frc_norm_dt[i]);
 
 		frc_integral += frc[i];
 		
@@ -524,7 +524,7 @@ QualityScores InterpolatedFRC::finish(const unsigned int num_particles)
 		
 		float adjusted_ssnr = tmp/((1.0-tmp)*num_particles);
 		normed_snr_integral += adjusted_ssnr;
-		snr_normed_frc_intergral += sqrtf(adjusted_ssnr/( 1.0 + adjusted_ssnr ));
+		snr_normed_frc_intergral += sqrt(adjusted_ssnr/( 1.0 + adjusted_ssnr ));
 	}
 	frc_integral /= size;
 	snr_normed_frc_intergral /= size;
